feat: Add file upload functionality for class materials with error handling
